fix blob_append_gsm7_packed : missing cast

+++ blob-emi.c
@@ -1176,14 +1176,14 @@ int blob_append_gsm7_packed(blob_t *out, const char *utf8, int unknown)
             return -1;
 
         if (c > 0xff) {
-            pack |= ((c >> 8) << (7 * septet));
+            pack |= ((uint64_t)(c >> 8) << (7 * septet));
             if (++septet == 8) {
                 blob_append_gsm_aligned_pack(out, pack, 7);
                 septet = 0;
                 pack = 0;
             }
         }
-        pack |= ((c & 0x7f) << (7 * septet));
+        pack |= ((uint64_t)(c & 0x7f) << (7 * septet));
         if (++septet == 8) {
             blob_append_gsm_aligned_pack(out, pack, 7);
             septet = 0;
